Fix yaml test ignore comments
